backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
pydantic-settings
pytest
httpx
sqlmodel
clickhouse-connect
redis
